# docs/swagger.yaml
basePath: /
definitions:
  handler.LoginRequest:
    properties:
      password:
        description: 用户密码
        type: string
      username:
        description: 用户姓名
        type: string
    required:
    - password
    - username
    type: object
  handler.RegisterRequest:
    properties:
      email:
        description: 用户邮箱
        type: string
      password:
        description: 用户密码
        maxLength: 50
        minLength: 6
        type: string
      username:
        description: 用户姓名
        maxLength: 50
        minLength: 2
        type: string
    required:
    - email
    - password
    - username
    type: object
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 响应数据
      message:
        description: 提示信息
        type: string
    type: object
info:
  contact: {}
  description: go-base project API documentation
  title: go-base API
  version: "1.0"
paths:
  /api/login:
    post:
      consumes:
      - application/json
      description: 用户名密码登录并返回 token
      parameters:
      - description: 登录参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 用户管理
  /api/register:
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 注册参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - 用户管理
  /api/userInfo:
    get:
      description: 需要携带 Bearer Token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"